// frontend/src/state/types/deployment.d.ts
import { Conversion, EnvironmentVar, Volume } from "./misc";

export type ProtoDeployment = {
  name: string;
  serverID: string;
  owner: string;

  buildID?: string; // if deploying a monitor build. can build / get build updates from the deployment page.
  image?: string; // this is set on creation using the attached buildID, or is passed directly
  latest?: boolean; // if custom image, use this to add :latest
  ports?: Conversion[];
  volumes?: Volume[];
  environment?: EnvironmentVar[];
  network?: string;
  restart?: string;
  postImage?: string;
  logTail?: number;
  containerUser?: string;
  logToAWS?: boolean;
  useServerRoot?: boolean;
  autoDeploy?: boolean;
  encryptedEnv?: string;
};

export interface Deployment extends ProtoDeployment {
  _id: string
  folderName: string
  containerName: string
}

export interface Deployments {
  [index: string]: Deployment
}